+++ aske-e.cabal
@@ -73,10 +73,10 @@ library
                         Language.ASKEE.Latex.Print
                         Language.ASKEE.Latex.Syntax
 
-                        Language.ASKEE.ModelStratify.GeoGraph
-                        Language.ASKEE.ModelStratify.Stratify
-                        Language.ASKEE.ModelStratify.Syntax
-                        Language.ASKEE.ModelStratify.Topology
+                        Language.ASKEE.AlgebraicJulia.GeoGraph
+                        Language.ASKEE.AlgebraicJulia.Stratify
+                        Language.ASKEE.AlgebraicJulia.Syntax
+                        Language.ASKEE.AlgebraicJulia.Topology
 
                         Language.ASKEE.RNet.GenLexer
                         Language.ASKEE.RNet.GenParser
@@ -97,7 +97,6 @@ library
                         Language.ASKEE.Model
                         Language.ASKEE.ModelType
                         Language.ASKEE.Panic
-                        Language.ASKEE.Simulate
                         Language.ASKEE.SimulatorGen
                         Language.ASKEE.Storage
