--- src/Language/ASKEE/Simulate.hs
@@ -1,13 +1,0 @@
-{-# LANGUAGE TemplateHaskell #-}
-module Language.ASKEE.Simulate where
-
--- import qualified Data.Map  as Map
--- import           Data.Map  ( Map )
--- import           Data.Text ( Text )
-
--- import           Language.ASKEE.DataSeries   ( DataSeries(..) )
--- import           Language.ASKEE.DEQ          ( loadDiffEqsFrom )
--- import qualified Language.ASKEE.DEQ.Simulate as Sim
--- import           Language.ASKEE.ModelType    ( ModelType )
--- import           Language.ASKEE.Storage      ( DataSource )
-
